Processes & Userspace (#11)
Rewrote the original process class and thread manager/scheduler
Each process gets its' own memory space
Setup switching userspace & handling userspace faults
Actual syscall setup for various syscalls
IPC Shared memory & Message passing
Processes can have args in userspace
GUI Panic handler
Spinlocks
Fixed reserve PMM function

+++ kernel/include/memory/memoryIO.h
@@ -17,9 +17,9 @@ namespace MaxOS{
          */
         class MemIO {
             protected:
-                uint32_t m_address;
+                uintptr_t m_address;
 
-                MemIO(uint32_t address);
+                MemIO(uintptr_t address);
                 ~MemIO();
             };
 
@@ -29,7 +29,7 @@ namespace MaxOS{
          */
         class MemIO8Bit : public MemIO {
             public:
-                MemIO8Bit(uint32_t address);
+                MemIO8Bit(uintptr_t address);
                 ~MemIO8Bit();
 
                 virtual void write(uint8_t data);
@@ -42,7 +42,7 @@ namespace MaxOS{
          */
         class MemIO16Bit : public MemIO {
             public:
-                MemIO16Bit(uint32_t address);
+                MemIO16Bit(uintptr_t address);
                 ~MemIO16Bit();
 
                 virtual void write(uint16_t data);
@@ -55,7 +55,7 @@ namespace MaxOS{
          */
         class MemIO32Bit : public MemIO {
             public:
-                MemIO32Bit(uint32_t address);
+                MemIO32Bit(uintptr_t address);
                 ~MemIO32Bit();
 
                 virtual void write(uint32_t data);
@@ -68,7 +68,7 @@ namespace MaxOS{
          */
         class MemIO64Bit : public MemIO {
         public:
-            MemIO64Bit(uint32_t address);
+            MemIO64Bit(uintptr_t address);
             ~MemIO64Bit();
 
             virtual void write(uint64_t data);
@@ -78,9 +78,9 @@ namespace MaxOS{
     }
 }
 
-void* memcpy(void* destination, const void* source, uint32_t num);
-void* memset(void* ptr, int value, uint32_t num);
-void* memmove(void* destination, const void* source, uint32_t num);
-int memcmp(const void* ptr1, const void* ptr2, uint32_t num);
+void* memcpy(void* destination, const void* source, uint64_t num);
+void* memset(void* ptr, int value, uint64_t num);
+void* memmove(void* destination, const void* source, uint64_t num);
+int memcmp(const void* ptr1, const void* ptr2, uint64_t num);
 
 #endif //MAXOS_MEMORY_MEMORYINPUTOUTPUT_H
